Started code clean-up/refactor
