+++ client/styles/GlobalStyles.js
@@ -1,0 +1,2 @@
+import { StyleSheet } from "react-native";
+
